+++ packages/secret-numbers/test/webpack.config.js
@@ -1,5 +1,5 @@
 "use strict";
-const { merge } = require("webpack-merge");
+const merge = require("webpack-merge");
 const { webpackForTest } = require("../../../webpack.test.config");
 const baseConfig = require("../webpack.base.config");