+++ packages/secret-numbers/webpack.config.js
@@ -1,6 +1,6 @@
 const path = require("path");
 const { getDefaultConfiguration, wrapForEnv } = require("../../webpack.config");
-const { merge } = require("webpack-merge");
+const merge = require("webpack-merge");
 
 module.exports = wrapForEnv(
   "xrplf-secret-numbers",