+++ packages/xrpl/HISTORY.md
@@ -4,6 +4,9 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ## Unreleased
 
+### Fixed
+* Updated the linter devDependencies to latest versions @xrplf/eslint-config@3.0.0, eslint-plugin-array-func@5.1.0 and eslint-plugin-tsdoc@0.5.0. As a consequence of the updates to `eslint-plugin-array-func`, all the eslint-config files have been updated from CommonJS format to ESM format to maintain compatibility.
+
 ## 4.5.0 (2025-12-16)
 
 ### Added