+++ packages/xrpl/HISTORY.md
@@ -6,6 +6,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ### Fixed
 * Updated the linter devDependencies to latest versions @xrplf/eslint-config@3.0.0, eslint-plugin-array-func@5.1.0 and eslint-plugin-tsdoc@0.5.0. As a consequence of the updates to `eslint-plugin-array-func`, all the eslint-config files have been updated from CommonJS format to ESM format to maintain compatibility.
+* Accomodated updates to the APIs of `webpack-merge`, `typescript v5.9` `@noble/curves` dependencies. No behavioral changes introduced.
 
 ### Added
 * Add `faucetProtocol` (http or https) option to `fundWallet` method. Makes `fundWallet` work with locally running faucet servers.