+++ packages/xrpl/HISTORY.md
@@ -7,6 +7,9 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 ### Added
 * Add `faucetProtocol` (http or https) option to `fundWallet` method. Makes `fundWallet` work with locally running faucet servers.
 
+### Fixed
+* Better error handling in the `Client` for edge case errors
+
 ## 4.5.0 (2025-12-16)
 
 ### Added